+++ cropclassification/postprocess/classification_reporting.py
@@ -1,6 +1,4 @@
-"""
-Module with some helper functions to report on the classification results.
-"""
+"""Module with some helper functions to report on the classification results."""
 
 import logging
 from pathlib import Path
@@ -38,16 +36,19 @@ def write_full_report(
     Args:
         parcel_predictions_geopath: File name of geofile with the parcels with their
             predictions.
-        parcel_train_path:
-        parcel_classification_data_path:
         output_report_txt: File name of txt file the report will be written to.
         parcel_ground_truth_path: List of parcels with ground truth to calculate
             eg. alfa and beta errors. If None, the part of the report that is based on
             this data is skipped
+        parcel_train_path: path to the file with the parcels that were used for
+            training.
+        parcel_classification_data_path: path to the file with the classification data.
+        force: If True, the report will be written even if the output file already
+            exists.
+
 
     TODO: refactor function to split logic more...
     """
-
     # If force == False Check and the output file exists already, stop.
     output_report_html = Path(str(output_report_txt).replace(".txt", ".html"))
     if force is False and output_report_txt.exists() and output_report_html.exists():
@@ -577,12 +578,13 @@ def write_full_report(
 
             # Pct BETA errors=beta errors/(beta errors + correct farmer declarations)
             beta_numerator_conclusions = ["FARMER-WRONG_PRED-DOESNT_OPPOSE:ERROR_BETA"]
-            beta_denominator_conclusions = (
-                df_parcel_gt[columnname]
-                .loc[df_parcel_gt[columnname].str.startswith("FARMER-CORRECT")]
+            beta_denominator_conclusions = list(beta_numerator_conclusions)
+            beta_denominator_conclusions.extend(
+                df_parcel_gt.loc[
+                    df_parcel_gt[columnname].str.startswith("FARMER-CORRECT")
+                ][columnname]
                 .unique()
                 .tolist()
-                + beta_numerator_conclusions
             )
             beta_numerator = len(
                 df_parcel_gt.loc[
@@ -607,20 +609,20 @@ def write_full_report(
             html_data["PREDICTION_QUALITY_BETA_TEXT"] += "<br/>" + message
 
             # Pct THETA errors=theta errors/(theta errors + correct farmer declarations)
-            theta_numerator_conclusions = (
+            theta_numerator_conclusions = list(
                 df_parcel_gt[columnname]
                 .loc[df_parcel_gt[columnname].str.startswith("FARMER-WRONG_PRED-DOUBT")]
                 .unique()
                 .tolist()
             )
-            theta_denominator_conclusions = (
+            theta_denominator_conclusions = list(theta_numerator_conclusions)
+            theta_denominator_conclusions.extend(
                 df_parcel_gt[columnname]
                 .loc[
                     df_parcel_gt[columnname].str.startswith("FARMER-CORRECT_PRED-DOUBT")
                 ]
                 .unique()
                 .tolist()
-                + theta_numerator_conclusions
             )
             theta_numerator = len(
                 df_parcel_gt.loc[
@@ -1142,7 +1144,6 @@ def write_full_report(
 
 def _get_confusion_matrix_ext(df_predict, prediction_column_to_use: str):
     """Returns a dataset with an extended confusion matrix."""
-
     classes = sorted(
         np.unique(
             np.append(
@@ -1214,8 +1215,7 @@ def _get_confusion_matrix_ext(df_predict, prediction_column_to_use: str):
 def _add_prediction_conclusion(
     in_df, new_columnname, prediction_column_to_use, detailed: bool
 ):
-    """
-    Calculate the "conclusions" for the predictions
+    """Calculate the "conclusions" for the predictions.
 
     REMARK: calculating it like this, using native pandas operations, is 300 times
             faster than using DataFrame.apply() with a function!!!
@@ -1310,7 +1310,6 @@ def _add_prediction_conclusion(
 
 def _add_gt_conclusions(in_df, prediction_column_to_use):
     """Add some columns with groundtruth conclusions."""
-
     # Add the new column with a fixed value first
     gt_vs_declared_column = f"gt_vs_input_{prediction_column_to_use}"
     gt_vs_prediction_column = f"gt_vs_prediction_{prediction_column_to_use}"
@@ -1501,9 +1500,7 @@ def _get_errors_per_column(
     include_cumulative_columns: bool,
     ascending: bool = True,
 ):
-    """
-    Calculates a detailed overview about the number of errors per group specified.
-    """
+    """Calculates a detailed overview about the number of errors per group specified."""
     # First filter on the parcels we need to calculate the pct alpha errors
     df_predquality_filtered = df_predquality[
         df_predquality[pred_quality_column].isin(error_codes_denominator)
@@ -1612,11 +1609,8 @@ def _write_OA_per_pixcount(
 ):
     """Write a report of the overall accuracy that parcels per pixcount get."""
     # If force == False Check and the output file exists already, stop.
-    if force is False and output_report_txt.exists():
-        logger.warning(
-            "collect_and_prepare_timeseries_data: output file already exists and "
-            f"force is False, so stop: {output_report_txt}"
-        )
+    if not force and output_report_txt.exists():
+        logger.warning(f"output file already exists so stop: {output_report_txt}")
         return
 
     # Write output...