+++ cropclassification/predict/__init__.py
@@ -1,0 +1,1 @@
+"""Pridict module for crop classification."""